add cmake versioning
to be capable of using pontos in cmake projects this creates a simple
cmake parser to get and update the version within CMakeLists.txt.

This parser is then used within a cmake version command.

This main script verifies if there is a CMakeLists.txt or a
pyproject.toml and decides based on that if it runs PoetryCommand or
CMakeCommand.