Skip running async based uinttests with Python 3.7
The IsolatedAsyncioTestCase has been introduced with Python 3.8.
Supporting the async tests would mean having to packport this test class
and also the async mocks. This seems to be to much effort to support a
Python version that gets only security fixes for less then a year from
now. Therefore just try to skip the async tests on Python 3.7.