Change: Refactor async GitHub API to return models
Instead of returning "JSON" data now the methods are returning fitting
model instances. By using models it is much easier to work with the API
because it is not necessary to lookup the provided data via print or
the docu. Now it is possible to use autocompletion of your editor.