Add: Allow to parse datetime values from dicts for Model instances
With this change datetime cann be used on Model classes. Currently we
just parse the string via fromisoformat without checking the timezone.
Not having a timezone is wrong in 99% percent of the cases. Therefore it
might be required that we just set UTC for datetimes parsed without a
timezone in future.