Change: Use code checkout to release pontos
When using our action to release pontos we actually use the last
released version of pontos to create the new release of pontos. If
pontos has a bug in the new release code we can't fix it without
uploading a new release to pypi manually with this approach. Therefore
drop the action usage and release pontos with it's current to be
released code. This ensures that the release fails if there is some bug
in the release code.