Change: Refactor ChangelogBuilder to allow empty releases
With this commit the API of the ChangelogBuilder did change. Now it is
possible to create a changelog without requiring to write it to a file
by adding a `create_changelog` method. The last version (formerly named
current version) and next version (also known as release version) are
now part of creating the change and not of the constructor anymore. This
allows to create different changelogs for several versions with the same
ChangelogBuilder instance.

Creating unreleased changelogs and links for the initial release are
fixed additionally.