Change: Remove terminal from ChangelogBuilder
Don't print output in ChangelogBuilder. Printing as a kind of UI should
only be done at the CLI side of the call stack. If something like this
is required it should be made available to the CLI code directly so that
the CLI can print the desired information. For example by using a
generator/iterator and yielding the information.