Change: Merge prepare into release command
Extend the release CLI to include the functionality of the prepare CLI.
With this change no intermediate prepare command is required to create a
release anymore.

Additionally no changelog file is written to the disk anymore.

To test the release it is possible to run release locally by adding the
`--local` argument where only commits and tags are created but not
pushed to a remote repo and no release is published at GitHub.