Improve version update (#708)
* Change: Update semantics of VersionUpdate

Clarify the meaning of None for previous property and add an is_update
method for formalizing if a VersionUpdate is actually an update or not.

* Change: Allow version comparisons against None

Allow to compare a Version instance to be compared against None. Don't
raise an exception in this case.

* Change: Update go command to return None if no previous version is found

Don't try to get the lastest version per git tag because this may fail
heavily when different release series are used. Therefore just return
None for the previous version when no current version could be
determined.