Change: Rework release process by using tags for last release version
Before this change the current release version was derived from the
project settings. Additionally the last release version was determined
from the git tags in a complicated way keeping the current release
version in mind to support different release series. The last release
version was used only to create the changelog for the release.

With this change getting the current version from the project settings
is dropped and only the last release version is used. Determining the
last release version from the git tags is simplified by just using the
latest tag (sorted by versions) and optionally a release series.

If a release series is provided the last tag fitting to the series is
used as last release version. A release series can be something like
"1.0", "2.0", "20.4" or even just "1".