Add: Extend conventional commits for dependency updates
Allow to include dependency updates in the conventional commits used for
generating release changelogs. Commit messages starting with `[Dd]eps`
are considered as conventional commits for dependency changes. They will
put into a new `Dependencies` section in the generated changelog.

This allows to include updates created by dependabot to be included in
the changelog by setting the `prefix` to `Deps` in the `commit-message`
setting of the dependabot config.