Fix: When updating a version ignore the current version in case of errors
During a version update the project's current version is parsed. This
may result in VersionErrors if for example the versioning scheme is
changed and the new versioning scheme doesn't recognize the current
version format (PEP440 based 22.4.1.dev1 can't be parsed as semver).

The current version is just parsed during the update to check if it is a
real update or if the same version is set. Therefore ignore the
VersionErrors during the version update and just set the new version.