Add: Introduce Version.parsed_version property
Introduce a property to get the original version string from which the
version has been parsed initially. This is helpful if the version has
been adjusted during parsing. For example PEP440 is very lax when
parsing a string but converts it to a canonicalized version. In this
case it might be necessary to have access to the original parsed version
string.