Fix: Fix creating changelog for non canonic versions
If the last tag used a non canonic version and it's parsed into a
version instance the version string might be adjusted. In that case it
wasn't possible to get the diff since the last release because the
version string doesn't fit to the git tag. Therefore use the
parsed_string now for getting the changelog since the last release.