Fix: Use "stable" URL for downloading GitHub release assets
Using the url of the assets json response returns different data with
every request. This URL is therefore unstable and every downloaded file
will have a different hash su,m. For being able to create signature we
need a stable API that returns the same data with every request.
Therefore don't use the url property from the json response and instead
the browser_download_url which is stable.