Add: Allow to enforce uploading releases as GitHub pre-releases (#894)
* Add: Allow to enforce uploading releases as GitHub pre-releases

When creating a release it is now possible to enforce uploading the
release with a pre-release status to GitHub.

* Fix typings and linting

* Change: Refactor tests for pontos-release create CLI

The CLI subcommand is intended to create a release and therefore got
renamed from `release` to `create`. Adjust the tests to reflect this
change.

* Add: Add tests for `--github-pre-release` flag for `pontos-release create`

Ensure the correct behavior of parsing the new flag.