Change: Introduce more flexible NVDResults class
Before all NVD API classes methods returned an async iterator. This
didn't allow much control of what the user actually wants and how the
requests are issued. To improve the situation a new NVDResults class is
returned which itself is an async iterator so that the previous API is
kept compatible.

But additionally the NVDResults instance allows to get the plain JSON
data, the number of available results and also to iterate over chunks of
results (which the NVD API is always returning).

Most important improvement the NVDResults instance keeps the state. That
means if an http error occurs it is possible to request the same data
again. With the old API the requests need to start from the beginning if
something did go wrong. For example if we downloaded already 100k CVEs
and a http error was raised we needed to start from CVE number 1 again.
With the new implementation we can just continue with the last request
again.