Change: Include changes from pre-releases in non pre-releases
Ignore pre-releases when determining the last release version if a non
pre-release is created. This includes all the changes of the previous
pre-releases into the changelog.

But still consider pre-releases when determining the last release when
a release series is used. Otherwise we might not determine a last
release when there are only pre-releases in the release series yet.