Add: delete_all on labels
Unlike the description says within set_all labels are not overriden but
labels that are not within the POST data but are on the PR stay.

For the use case that an action wants to override managed labels it
needs the possibility to delete previous labels.

For that the delete_all method is introduced.