Change: Rust support for workspace.package.version
This is a BREAKING CHANGE to all of those Rust projects using Pontos,
if it is used to keep all versions of the member crates in sync with the
root crate version.

This is discuraged in favor of configuring `[workspace.package.version]`
in the root Cargo.toml and `version.workspace = true` in the member
crates Cargo.toml to allow for unified metadata like the version.

Also Pontos now correctly handles [workspace.package.version] and
[package.version]. The latter always has higher priority, if both
version properties are set.